fix mv error on container restart

+++ docker-run.sh
@@ -1,5 +1,7 @@
 #!/bin/sh
 
-mv /etc/nginx/conf.d/default.conf /etc/nginx/default.conf.bak
+if [ -f /etc/nginx/conf.d/default.conf ] ; then
+	mv /etc/nginx/conf.d/default.conf /etc/nginx/default.conf.bak
+fi
 uwsgi /etc/uwsgi/ecss.ini &
 exec nginx -g 'daemon off;'
